Move sampleMatrixRecords to GridMatrix+Sample.swift
